+++ src/oem_logistic_sparse.h
@@ -214,7 +214,7 @@ protected:
         {
             SpMat XXtmp(nvars, nvars);
             
-            int numrowscurfirst = floor(nobs / ncores);
+            int numrowscurfirst = std::floor(double(nobs) / double(ncores));
             
             #pragma omp parallel
             {
@@ -226,10 +226,9 @@ protected:
                 #pragma omp for schedule(static) nowait
                 for (int ff = 0; ff < ncores; ++ff)
                 {
-                    
                     if (ff + 1 == ncores)
                     {
-                        int numrowscur = nobs - (ncores - 1) * floor(nobs / ncores);
+                        int numrowscur = nobs - (ncores - 1) * std::floor(double(nobs) / double(ncores));
                         
                         XXtmp_private += SpMat(nvars, nvars).selfadjointView<Upper>().
                         rankUpdate(X.bottomRows(numrowscur).adjoint() * 
@@ -667,7 +666,7 @@ public:
                     }
                 } else 
                 {
-                    int numrowscur = floor(nobs / ncores);
+                    int numrowscur = std::floor(double(nobs) / double(ncores));
                     int numrowscurfirst = numrowscur;
                     
                     #pragma omp parallel for schedule(static)
@@ -676,7 +675,7 @@ public:
                         
                         if (ff + 1 == ncores)
                         {
-                            numrowscur = nobs - (ncores - 1) * floor(nobs / ncores);
+                            numrowscur = nobs - (ncores - 1) * std::floor(double(nobs) / double(ncores));
                             if (intercept)
                             {
                                 if (standardize)
